Relocate vm folder under models
