+++ Garage_2_0/Views/_ViewImports.cshtml
@@ -1,6 +1,6 @@
 ﻿@using Garage_2_0.Models
 @using Garage_2_0.Models.Enums
-@using Garage_2_0.ViewModels
+@using Garage_2_0.Models.ViewModels
 
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
 @addTagHelper *, garage-2.0
